third_party/WebKit/LayoutTests/shadow-dom/crashes/link-style-change-href-assert.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="host">This text should not be red and there should be no asserts.</div>
<script>
    var root = host.attachShadow({mode:'open'});
    root.innerHTML = '<slot/><link rel="stylesheet" href="data:text/css,:host{color:red}">';

    test(() => {
        assert_equals(getComputedStyle(host).color, "rgb(255, 0, 0)", "Host color is red.");
    }, "Check initial color from applied stylesheet.");

    test(() => {
        root.querySelector("link").setAttribute("href", "data:text/css,");
        assert_equals(getComputedStyle(host).color, "rgb(0, 0, 0)", "Host color is black after href is changed.");
    }, "Changing href on a link in shadow should not trigger an assert.");
</script>
